Add custom event handling API
Custom event handling API will allow custom behavior as well as registering event handlers at different locations to listen in on custom events.

The only event available from the JavaFX API is the ActionEvent, but that must be handled within the class emitting the event, and is also to too general to listen in to.

The structure written better fits the current design.