Create API for a swappable panel
A swappable pane can be created by implementing the swappable interface, where it can define it's replacement UiPane.

The SwitchPaneEventHandler enforces that the Swappable provided to it is a pane as well, and that parent provided to it is indeed a parent of the Swappable.

The handler then executes the swap and modifies the JavaFX node tree.