Add minimum integration for PF4J and ZeroMQ
This commit adds ZeroMQ as a plugin for PF4J running on a separate thread than the main application, to prevent the ZeroMQ receiving message loop from blocking the main UI thread.

This commit also adds ZeroMQ as an extension of the Pf4jMessagable interface, allowing configurations such as adding different messaging services that handle different kinds of data.