Add Model and ZeroMQ data handling
This commit adds a Drone model to store Drone telemetry, and makes use of Spring events to publish data from plugin extensions to data handling services. This decouples the plugin functionalities from the data handling classes.

Future improvements can include changes to the ZeroMQ plugin so that lifecycle methods can be improved to properly end threads as well as close the ZeroMQ context.