Confer model changes to UI
DroneModel which is the handler class for the model, will propagate events that will be listened to by Ui classes. These Ui classes will then update the Ui based on the data passed to them by the DroneModel through the events.

Because only the JavaFX thread can update JavaFX UI components, and the application is multithreaded, any change to Ui components must be done through the JavaFX Platform.runLater method.