Add framework for reliable request-reply messaging
This commit adds ZeroMQ plugin functionality to send commands via reliable request-reply functionality. This means that changes will not be propagated and registered until a reply is received by DJIAAPP.

This commit also makes use of the visitor pattern in the last commit to process outgoing DroneCommandMessages and incoming DroneCommandReplyMessages. This way, the DroneMessageService acts as a dispatcher of sorts, providing logic based on the messages it receives.

The last change is to provide a framework for propagating changes to the model based on the status of the CommandReply. The model will then release events that Ui classes will listen to to update the UI components.