Implement Switching to Drone Feed from Map View
Clicking the small Drone Feed Window will switch from the Map View to the Video View.

UiMapBasePane now extends SwappableUiPane that allows it to get the pane to swap to. It's parent now contains an eventHandler that swaps between UiMapBasePane and UiVideoTab once the Mini Drone Feed in the Map View is clicked.